notificaiton controller

+++ app/Http/Controllers/NotificationController.php
@@ -2,31 +2,52 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Resources\NotificationResource;
 use Illuminate\Support\facades\Auth;
 use Illuminate\Http\Request;
+use App\Models\User;
+use App\Notifications\NewReleaseNotification;
+use Illuminate\Support\Facades\Gate;
+use Illuminate\Notifications\Notification;
 
 class NotificationController extends Controller
 {
-    public function markAsRead($id)
+
+    public function markAsRead(Notification $id)
     {
-        Auth::user()->markAllNotificationsAsRead($id);
-        return response()->json(['message' => 'Notification marked as read.']);
-    }
-    public function markAllAsRead()
-    {
+        Gate::authorize('check', $id);
+        // $user = Auth::user();
+        // if (!$user) {
+        //     return new NotificationResource(['message' => 'User not authenticated.'], 401);
+        // }
+        // $notification = Auth::user()->notifications()->find($id);
+        // if ($notification) {
+        //     $notification->markAsRead();
+        //     return new NotificationResource(['message' => 'Uthorized.'], 403);
+        // }
         Auth::user()->markAllNotificationsAsRead();
-        return response()->json(['message' => 'All notifications marked as read.']);
+        return new NotificationResource(['message' => 'Notification mark as read.'], 401);
     }
 
-    public function markAsUnread($id)
+
+    public function markAllAsRead(Notification $id)
     {
+        Gate::authorize('modify', $id);
+        Auth::user()->markAllNotificationsAsRead();
+        return new NotificationResource(['message' => 'Notification mark all as read.'], 401);
+    }
+
+    public function markAsUnread(Notification $id)
+    {
+        Gate::authorize('modify', $id);
         Auth::user()->markNotificationAsUnread($id);
-        return response()->json(['message' => 'Notification marked as unread.']);
+        return new NotificationResource(['message' => 'Notification mark all as unread.'], 401);
     }
 
-    public function deleteNotification($id)
+    public function deleteNotification(Notification $id)
     {
+        Gate::authorize('modify', $id);
         Auth::user()->deleteNotification($id);
-        return response()->json(['message' => 'Notification deleted.']);
+        return new NotificationResource(['message' => 'Notification deleted.'], 401);
     }
 }
